NLU Parser for English commands in Maverick domain.
The parser is based on CFG rules. It accepts sentences for messaging that 1) start with a slot such as the sending time, a contact slot, a command verb, an intent or initial/polite talk and 2) end with the message content/body. Every slot is recognized by a triggering keyword such as "at" for the time, "to" for the contact, and "content" or "contains" for the content slot.